+++ src/common.jl
@@ -130,124 +130,6 @@ function class_statistics!(M::AbstractMatrix, nₘ::Vector{<:Integer}, X::Abstra
 end
 
 
-### Model Parameters
-
-mutable struct DiscriminantParameters{T}
-    "Model fit indicator - `true` if model has been fit"
-    fit::Bool
-    "Dimension along which observations are stored (1 for rows, 2 for columns)"
-    dims::Int
-    "Number of classes"
-    m::Int
-    "Number of predictors"
-    p::Int
-    "Matrix of class centroids (one per row or column - see `dims`)"
-    M::Matrix{T}
-    "Prior-weighted overall centroid"
-    μ::Vector{T}
-    "Vector of class prior probabilities"
-    π::Vector{T}
-    "Vector of Class counts"
-    nₘ::Vector{Int}
-    "Shrinkage parameter"
-    γ::Union{Nothing,T}
-    "Overall covariance matrix after `γ` regularization"
-    Σ::Union{Nothing,Matrix{T}}
-    "Determinant of the overall covariance matrix"
-    detΣ::Union{Nothing,T}
-    function DiscriminantParameters{T}() where T
-        new{T}(false, 0, 0, 0, Matrix{T}(undef,0,0), Vector{T}(undef,0), Vector{T}(undef,0),
-               Vector{Int}(undef,0), nothing, nothing, zero(T))
-    end
-end
-
-function set_dimensions!(Θ::DiscriminantParameters{T},
-                         y::Vector{<:Integer},
-                         X::Matrix{T},
-                         dims::Integer) where {T}
-    n, p = check_dims(X, dims=dims)
-    m = maximum(y)
-    
-    n₂ = length(y)
-    n₂ == n || throw(DimensionMismatch("observation count along length of class index " *
-                                       "vector y must match dimension $(dims) of data " *
-                                       "matrix X (got $(n₂) and $(n))"))
-
-    Θ.dims = dims
-    Θ.m = m
-    Θ.p = p
-
-    return Θ
-end
-
-function set_gamma!(Θ::DiscriminantParameters{T}, γ::Union{Nothing,Real}) where T
-    if γ !== nothing
-        0 ≤ γ ≤ 1 || throw(DomainError(γ, "γ must be in the interval [0,1]"))
-    end
-    Θ.γ = γ
-
-    return Θ
-end
-
-function set_statistics!(Θ::DiscriminantParameters{T},
-                         y::Vector{<:Integer},
-                         X::Matrix{T},
-                         centroids::Union{Nothing,AbstractMatrix}) where T
-    dims = Θ.dims
-    m = Θ.m
-    p = Θ.p
-
-    # Compute centroids and class counts from data if not specified
-    if centroids === nothing
-        Θ.M = dims == 1 ? Matrix{T}(undef, m, p) : Matrix{T}(undef, p, m)
-        Θ.nₘ = Vector{Int}(undef, m)
-
-        class_statistics!(Θ.M, Θ.nₘ, X, y, dims=dims)
-    else
-        m₂, p₂ = check_dims(centroids, dims=dims)
-        if m₂ != m
-            throw(DimensionMismatch("class count along dimension $(dims) of centroid " * 
-                                    "matrix M must match maximum class index found in " *
-                                    "class index vector y (got $(m₂) and $(m))"))
-        elseif p₂ != p
-            altdims = dims == 1 ? 2 : 1
-            throw(DimensionMismatch("predictor count along dimension $(altdims) of " *
-                                    "centroid matrix M must match dimension $(dims) of " *
-                                    "data matrix X (got $(p₂) and $(p))"))
-        end
-
-        Θ.M = copyto!(similar(centroids, T), centroids)
-        Θ.nₘ = class_counts!(Vector{Int}(undef, m), y)
-    end
-
-    validate_class_counts(Θ.nₘ)
-
-    return Θ
-end
-
-function set_priors!(Θ::DiscriminantParameters{T}, 
-                     π::Union{Nothing,AbstractVector}) where T
-    m = Θ.m
-
-    # Compute priors from class frequencies in data if not specified
-    if π === nothing
-        Θ.π = broadcast!(/, Vector{T}(undef, m), Θ.nₘ, sum(Θ.nₘ))
-    else
-        m₂ = length(π)
-        if m₂ != m
-            throw(DimensionMismatch("class count along length of class prior probability " *
-                                    "vector π must match maximum class index found in " *
-                                    "class index vector y (got $(m₂) and $(m))"))
-        end
-        validate_priors(π)
-
-        Θ.π = copyto!(similar(π, T), π)
-    end
-
-    return Θ
-end
-
-
 ### Regularization
 
 """
@@ -288,106 +170,3 @@ function regularize!(Σ::AbstractMatrix{T}, γ::T) where {T}
 
     return Σ
 end
-
-
-### Data Whitening Functions
-
-"""
-    whiten_data!(X; dims, df)
-
-Compute a whitening transform matrix for centered data matrix `X`. Use `dims=1` for 
-row-based observations and `dims=2` for column-based observations. The `df` parameter 
-specifies the effective degrees of freedom.
-"""
-function whiten_data!(X::Matrix{T}; dims::Integer, df::Integer=size(X,dims)-1) where T
-    df > 0 || error("degrees of freedom must be greater than 0")
-
-    n, p = check_dims(X, dims=dims)
-
-    n > p || error("insufficient number of within-class observations to produce a full " *
-                   "rank covariance matrix ($(n) observations, $(p) predictors)")
-
-    if dims == 1
-        # X = QR ⟹ S = XᵀX = RᵀR
-        R = UpperTriangular(qr!(X, Val(false)).R)  
-    else
-        # Xᵀ = LQ ⟹ S = XXᵀ = LLᵀ = RᵀR
-        R = UpperTriangular(transpose(lq!(X).L))  
-    end
-
-    broadcast!(/, R, R, √(df))
-
-    detΣ = det(R)^2
-
-    W = try
-        inv(R)
-    catch err
-        if isa(err, LAPACKException) || isa(err, SingularException)
-            if err.info ≥ 1
-                error("rank deficiency detected (collinearity in predictors)")
-            end
-        end
-        throw(err)
-    end
-
-    if dims == 1
-        return (W, detΣ)
-    else
-        return (copy(transpose(W)), detΣ)
-    end
-end
-
-@inline regularize(x, y, γ) = (1-γ)*x + γ*y
-
-function whiten_data!(X::Matrix{T}, γ::T; dims::Integer, df::Integer=size(X,dims)-1) where T
-    n, p = check_dims(X, dims=dims)
-    
-    n > p || error("insufficient number of within-class observations to produce a full " *
-                   "rank covariance matrix ($(n) observations, $(p) predictors)")
-    
-    0 ≤ γ ≤ 1 || throw(DomainError(γ, "γ must be in the interval [0,1]"))
-
-    tol = eps(one(T))*p*maximum(X)
-
-    UDVᵀ = svd!(X, full=false)
-
-    D = UDVᵀ.S
-
-    if γ ≠ zero(T)
-        # Regularize: Σ = VD²Vᵀ ⟹ Σ(γ) = V((1-γ)D² + (γ/p)trace(D²)I)Vᵀ
-        broadcast!(σᵢ -> abs2(σᵢ)/df, D, D)  # Convert data singular values to Σ eigenvalues
-        broadcast!(regularize, D, D, mean(D), γ)
-        detΣ = prod(D)
-        broadcast!(√, D, D)
-    else
-        detΣ = prod(σᵢ -> abs2(σᵢ)/df, D)
-        broadcast!(/, D, D, √(df))
-    end
-
-    all(D .> tol) || error("rank deficiency (collinearity) detected with tolerance $(tol)")
-
-    # Whitening matrix
-    if dims == 1
-        Vᵀ = UDVᵀ.Vt
-        Wᵀ = broadcast!(/, Vᵀ, Vᵀ, D)  # in-place diagonal matrix multiply DVᵀ
-    else
-        U = UDVᵀ.U
-        Wᵀ = broadcast!(/, U, U, transpose(D))
-    end
-
-    return (copy(transpose(Wᵀ)), detΣ)
-end
-
-
-function whiten_cov!(Σ::AbstractMatrix{T}, γ::T=zero(T)) where T
-    (p = size(Σ, 1)) == size(Σ, 2) || throw(DimensionMismatch("Σ must be square"))
-
-    0 ≤ γ ≤ 1 || throw(DomainError(γ, "γ must be in the interval [0,1]"))
-    
-    if γ != 0
-        regularize!(Σ, γ)
-    end
-    
-    W = inv(cholesky!(Σ, Val(false); check=true).U)
-end
-
