+++ src/quadratic.jl
